refactor: organización de ViewModels en el paquete bovinos y limpieza de imports
