+++ terrabit_app/app/src/main/java/com/example/terrabit_app/MainActivity.kt
@@ -15,8 +15,8 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.terrabit_app.ui.navigation.Navigation
 import com.example.terrabit_app.ui.theme.Terrabit_appTheme
-import com.example.terrabit_app.viewmodel.ConfigurationViewModel
-import com.example.terrabit_app.viewmodel.DrawerViewModel
+import com.example.terrabit_app.viewmodel.bovinos.ConfigurationViewModel
+import com.example.terrabit_app.viewmodel.bovinos.DrawerViewModel
 import androidx.appcompat.app.AppCompatActivity
 import com.example.terrabit_app.utils.bluetooth.BluetoothViewModel
 import dagger.hilt.android.AndroidEntryPoint
