+++ terrabit_app/app/src/main/java/com/example/terrabit_app/ui/screen/DrawerScreen.kt
@@ -30,7 +30,7 @@ import com.example.terrabit_app.ui.theme.MainGreen
 import com.example.terrabit_app.ui.theme.MainOrange
 import com.example.terrabit_app.utils.UserPreferences
 import com.example.terrabit_app.utils.bluetooth.BluetoothViewModel
-import com.example.terrabit_app.viewmodel.DrawerViewModel
+import com.example.terrabit_app.viewmodel.bovinos.DrawerViewModel
 import kotlinx.coroutines.launch
 
 @RequiresApi(Build.VERSION_CODES.O)
