--- terrabit_app/app/src/main/java/com/example/terrabit_app/viewmodel/MainViewmodel.kt
@@ -1,26 +1,0 @@
-package com.example.terrabit_app.viewmodel
-
-import android.app.Application
-import android.util.Log
-import androidx.lifecycle.AndroidViewModel
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.example.terrabit_app.data.network.Identificadores.Identificadores
-import com.example.terrabit_app.data.network.Repositorio
-import com.example.terrabit_app.data.network.animales.RegistroMuerteBovi
-import com.example.terrabit_app.data.network.material.PetSolicitudMaterial
-import com.example.terrabit_app.data.network.material.Unitat
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
-import java.util.*
-
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class MainViewmodel @Inject constructor() : ViewModel(){
-
-}
