+++ terrabit_app/app/src/main/java/com/example/terrabit_app/viewmodel/bovinos/CorrecionSexoViewModel.kt
@@ -1,10 +1,9 @@
-package com.example.terrabit_app.viewmodel
+package com.example.terrabit_app.viewmodel.bovinos
 
 import android.util.Log
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.terrabit_app.data.Borrador
 import com.example.terrabit_app.data.network.Repositorio
 import com.example.terrabit_app.data.network.animales.PetModicarAnimal
 import com.example.terrabit_app.data.network.lista_bovinos.Animal
@@ -23,7 +22,9 @@ import com.example.terrabit_app.data.local.dao.BorradorDao
 import com.example.terrabit_app.data.local.dao.HistorialDao
 import com.example.terrabit_app.data.local.database.BorradorEntity
 import com.example.terrabit_app.data.local.database.HistorialEntity
-import com.example.terrabit_app.data.local.database.toBorrador
+import com.google.gson.reflect.TypeToken
+import java.io.IOException
+import java.net.SocketTimeoutException
 import java.util.UUID
 
 @HiltViewModel
@@ -163,7 +164,7 @@ class CorrecionSexoViewModel @Inject constructor(
                 borradorSesionId = borrador.id
                 val datos: Map<String, Any?> = Gson().fromJson(
                     borrador.datos,
-                    object : com.google.gson.reflect.TypeToken<Map<String, Any?>>() {}.type
+                    object : TypeToken<Map<String, Any?>>() {}.type
                 )
                 _identificadorCorreccionSexo.value = datos["identificador"] as? String ?: ""
                 _sexoCorreccionSeleccionado.value = datos["sexoSeleccionado"] as? String ?: ""
@@ -197,7 +198,7 @@ class CorrecionSexoViewModel @Inject constructor(
                 borradorSesionId = borrador.id
                 val datos: Map<String, Any?> = Gson().fromJson(
                     borrador.datos,
-                    object : com.google.gson.reflect.TypeToken<Map<String, Any?>>() {}.type
+                    object : TypeToken<Map<String, Any?>>() {}.type
                 )
                 _identificadorCorreccionSexo.value = datos["identificador"] as? String ?: ""
                 _sexoCorreccionSeleccionado.value = datos["sexoSeleccionado"] as? String ?: ""
@@ -285,13 +286,13 @@ class CorrecionSexoViewModel @Inject constructor(
                         }
                     }
                 }
-            } catch (e: java.net.SocketTimeoutException) {
+            } catch (e: SocketTimeoutException) {
                 withContext(Dispatchers.Main) {
                     _estadoCarga.value = false; _correccionSexoExitosa.value = false
                     _mensajeErrorCorreccionSexo.value = "Tiempo de espera agotado. La operación puede haberse completado, por favor verifique."
                     Log.e("Error Corrección Sexo", "Timeout: ${e.message}", e)
                 }
-            } catch (e: java.io.IOException) {
+            } catch (e: IOException) {
                 withContext(Dispatchers.Main) {
                     _estadoCarga.value = false; _correccionSexoExitosa.value = false
                     _mensajeErrorCorreccionSexo.value = "Error de conexión. Verifique su conexión a internet."
@@ -330,7 +331,7 @@ class CorrecionSexoViewModel @Inject constructor(
                     "codigoSexo" to codigoSexo
                 )
                 historialDao.insert(HistorialEntity(
-                    id = java.util.UUID.randomUUID().toString(),
+                    id = UUID.randomUUID().toString(),
                     tipo = "CORRECCION_SEXO",
                     fecha = SimpleDateFormat("dd/MM/yyyy", Locale.getDefault()).format(Date()),
                     hora = SimpleDateFormat("HH:mm", Locale.getDefault()).format(Date()),
@@ -349,7 +350,7 @@ class CorrecionSexoViewModel @Inject constructor(
                 val registro = historialDao.getAll().find { it.id == id } ?: return@launch
                 val datos: Map<String, Any?> = Gson().fromJson(
                     registro.datos,
-                    object : com.google.gson.reflect.TypeToken<Map<String, Any?>>() {}.type
+                    object : TypeToken<Map<String, Any?>>() {}.type
                 )
                 _identificadorCorreccionSexo.value = datos["identificador"] as? String ?: ""
                 _sexoCorreccionSeleccionado.value = datos["sexoSeleccionado"] as? String ?: ""
