+++ terrabit_app/app/src/main/java/com/example/terrabit_app/viewmodel/bovinos/IdentificacionAplazaViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.terrabit_app.viewmodel
+package com.example.terrabit_app.viewmodel.bovinos
 
 import android.util.Log
 import androidx.lifecycle.MutableLiveData
@@ -22,7 +22,9 @@ import com.example.terrabit_app.data.local.dao.BorradorDao
 import com.example.terrabit_app.data.local.dao.HistorialDao
 import com.example.terrabit_app.data.local.database.BorradorEntity
 import com.example.terrabit_app.data.local.database.HistorialEntity
-import com.example.terrabit_app.data.local.database.toBorrador
+import com.google.gson.reflect.TypeToken
+import java.io.IOException
+import java.net.SocketTimeoutException
 
 @HiltViewModel
 class IdentificacionAplazaViewModel @Inject constructor(
@@ -155,7 +157,7 @@ class IdentificacionAplazaViewModel @Inject constructor(
                 borradorSesionId = borrador.id
                 val datos: Map<String, Any?> = Gson().fromJson(
                     borrador.datos,
-                    object : com.google.gson.reflect.TypeToken<Map<String, Any?>>() {}.type
+                    object : TypeToken<Map<String, Any?>>() {}.type
                 )
                 _identificadorAnimal.value = datos["identificador"] as? String ?: ""
                 _fechaIdentificacion.value = datos["fechaIdentificacion"] as? String ?: ""
@@ -259,13 +261,13 @@ class IdentificacionAplazaViewModel @Inject constructor(
                         }
                     }
                 }
-            } catch (e: java.net.SocketTimeoutException) {
+            } catch (e: SocketTimeoutException) {
                 withContext(Dispatchers.Main) {
                     _estadoCarga.value = false; _identificacionExitosa.value = false
                     _mensajeErrorIdentificacion.value = "Tiempo de espera agotado. La operación puede haberse completado, por favor verifique."
                     Log.e("Error Corrección Identificacion", "Timeout: ${e.message}", e)
                 }
-            } catch (e: java.io.IOException) {
+            } catch (e: IOException) {
                 withContext(Dispatchers.Main) {
                     _estadoCarga.value = false; _identificacionExitosa.value = false
                     _mensajeErrorIdentificacion.value = "Error de conexión. Verifique su conexión a internet."
@@ -302,7 +304,7 @@ class IdentificacionAplazaViewModel @Inject constructor(
                     "fechaIdentificacion" to _fechaIdentificacion.value
                 )
                 historialDao.insert(HistorialEntity(
-                    id = java.util.UUID.randomUUID().toString(),
+                    id = UUID.randomUUID().toString(),
                     tipo = "IDENTIFICACION_APLAZADA",
                     fecha = SimpleDateFormat("dd/MM/yyyy", Locale.getDefault()).format(Date()),
                     hora = SimpleDateFormat("HH:mm", Locale.getDefault()).format(Date()),
@@ -321,7 +323,7 @@ class IdentificacionAplazaViewModel @Inject constructor(
                 val registro = historialDao.getAll().find { it.id == id } ?: return@launch
                 val datos: Map<String, Any?> = Gson().fromJson(
                     registro.datos,
-                    object : com.google.gson.reflect.TypeToken<Map<String, Any?>>() {}.type
+                    object : TypeToken<Map<String, Any?>>() {}.type
                 )
                 _identificadorAnimal.value = datos["identificador"] as? String ?: ""
                 _fechaIdentificacion.value = datos["fechaIdentificacion"] as? String ?: ""
