+++ terrabit_app/app/src/main/java/com/example/terrabit_app/viewmodel/bovinos/MainViewmodel.kt
@@ -1,0 +1,11 @@
+package com.example.terrabit_app.viewmodel.bovinos
+
+import androidx.lifecycle.ViewModel
+
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class MainViewmodel @Inject constructor() : ViewModel(){
+
+}
