+++ terrabit_app/app/src/main/java/com/example/terrabit_app/viewmodel/bovinos/MaterialDuplicadoViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.terrabit_app.viewmodel
+package com.example.terrabit_app.viewmodel.bovinos
 
 import android.util.Log
 import androidx.lifecycle.MutableLiveData
@@ -15,10 +15,13 @@ import com.example.terrabit_app.data.network.material.PetSolicitudDuplicado
 import com.example.terrabit_app.data.network.respuestas.RespuestaUnificada
 import com.example.terrabit_app.utils.UserPreferences
 import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
+import java.io.IOException
+import java.net.SocketTimeoutException
 import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
@@ -184,7 +187,7 @@ class MaterialDuplicadoViewModel @Inject constructor(
                 borradorSesionId = borrador.id
                 val datos: Map<String, Any?> = Gson().fromJson(
                     borrador.datos,
-                    object : com.google.gson.reflect.TypeToken<Map<String, Any?>>() {}.type
+                    object : TypeToken<Map<String, Any?>>() {}.type
                 )
                 _empresaSubministradora.value = datos["empresaSubministradora"] as? String ?: ""
                 codigoEmpresaSubministradora = datos["codigoEmpresaSubministradora"] as? String ?: ""
@@ -399,9 +402,9 @@ class MaterialDuplicadoViewModel @Inject constructor(
                         else -> { _registroExitoso.value = false; _mensajeError.value = "Error: Respuesta vacía del servidor" }
                     }
                 }
-            } catch (e: java.net.SocketTimeoutException) {
+            } catch (e: SocketTimeoutException) {
                 withContext(Dispatchers.Main) { _cargando.value = false; _registroExitoso.value = false; _mensajeError.value = "Tiempo de espera agotado." }
-            } catch (e: java.io.IOException) {
+            } catch (e: IOException) {
                 withContext(Dispatchers.Main) { _cargando.value = false; _registroExitoso.value = false; _mensajeError.value = "Error de conexión." }
             } catch (e: Exception) {
                 withContext(Dispatchers.Main) { _cargando.value = false; _registroExitoso.value = false; _mensajeError.value = "Error inesperado: ${e.message ?: "Error desconocido"}" }
@@ -443,7 +446,7 @@ class MaterialDuplicadoViewModel @Inject constructor(
                     "listaAnimales" to _listaAnimales.value
                 )
                 historialDao.insert(HistorialEntity(
-                    id = java.util.UUID.randomUUID().toString(),
+                    id = UUID.randomUUID().toString(),
                     tipo = "MATERIAL_DUPLICADO",
                     fecha = SimpleDateFormat("dd/MM/yyyy", Locale.getDefault()).format(Date()),
                     hora = SimpleDateFormat("HH:mm", Locale.getDefault()).format(Date()),
@@ -462,7 +465,7 @@ class MaterialDuplicadoViewModel @Inject constructor(
                 val registro = historialDao.getAll().find { it.id == id } ?: return@launch
                 val datos: Map<String, Any?> = Gson().fromJson(
                     registro.datos,
-                    object : com.google.gson.reflect.TypeToken<Map<String, Any?>>() {}.type
+                    object : TypeToken<Map<String, Any?>>() {}.type
                 )
                 _empresaSubministradora.value = datos["empresaSubministradora"] as? String ?: ""
                 codigoEmpresaSubministradora = datos["codigoEmpresaSubministradora"] as? String ?: ""
