+++ terrabit_app/app/src/main/java/com/example/terrabit_app/viewmodel/bovinos/ViewModelMuerteBovi.kt
@@ -1,11 +1,9 @@
-package com.example.terrabit_app.viewmodel
+package com.example.terrabit_app.viewmodel.bovinos
 
 import android.util.Log
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.terrabit_app.data.Borrador
-import com.example.terrabit_app.data.SharedPreferencesManager
 import com.example.terrabit_app.data.network.Repositorio
 import com.example.terrabit_app.data.network.animales.RegistroMuerteBovi
 import com.example.terrabit_app.data.network.lista_bovinos.Animal
@@ -24,7 +22,9 @@ import com.example.terrabit_app.data.local.dao.BorradorDao
 import com.example.terrabit_app.data.local.dao.HistorialDao
 import com.example.terrabit_app.data.local.database.BorradorEntity
 import com.example.terrabit_app.data.local.database.HistorialEntity
-import com.example.terrabit_app.data.local.database.toBorrador
+import com.google.gson.reflect.TypeToken
+import java.io.IOException
+import java.net.SocketTimeoutException
 
 @HiltViewModel
 class ViewModelMuerteBovi @Inject constructor(
@@ -190,7 +190,7 @@ class ViewModelMuerteBovi @Inject constructor(
                 borradorSesionId = borrador.id
                 val datos: Map<String, Any?> = Gson().fromJson(
                     borrador.datos,
-                    object : com.google.gson.reflect.TypeToken<Map<String, Any?>>() {}.type
+                    object : TypeToken<Map<String, Any?>>() {}.type
                 )
                 _tipoMuerte.value = datos["tipo"] as? String ?: ""
                 _codigoTipoMuerte.value = datos["codigoTipo"] as? String ?: ""
@@ -362,13 +362,13 @@ class ViewModelMuerteBovi @Inject constructor(
                         }
                     }
                 }
-            } catch (e: java.net.SocketTimeoutException) {
+            } catch (e: SocketTimeoutException) {
                 withContext(Dispatchers.Main) {
                     _cargandoMuerte.value = false; _registroMuerteExitoso.value = false
                     _mensajeErrorMuerte.value = "Tiempo de espera agotado. La operación puede haberse completado, por favor verifique."
                     Log.e("Error Registro Muerte", "Timeout: ${e.message}", e)
                 }
-            } catch (e: java.io.IOException) {
+            } catch (e: IOException) {
                 withContext(Dispatchers.Main) {
                     _cargandoMuerte.value = false; _registroMuerteExitoso.value = false
                     _mensajeErrorMuerte.value = "Error de conexión. Verifique su conexión a internet."
@@ -400,7 +400,7 @@ class ViewModelMuerteBovi @Inject constructor(
                     "coordenadaY" to _coordenadaY.value
                 )
                 historialDao.insert(HistorialEntity(
-                    id = java.util.UUID.randomUUID().toString(),
+                    id = UUID.randomUUID().toString(),
                     tipo = "MUERTE",
                     fecha = SimpleDateFormat("dd/MM/yyyy", Locale.getDefault()).format(Date()),
                     hora = SimpleDateFormat("HH:mm", Locale.getDefault()).format(Date()),
@@ -420,7 +420,7 @@ class ViewModelMuerteBovi @Inject constructor(
                 val registro = historialDao.getAll().find { it.id == id } ?: return@launch
                 val datos: Map<String, Any?> = Gson().fromJson(
                     registro.datos,
-                    object : com.google.gson.reflect.TypeToken<Map<String, Any?>>() {}.type
+                    object : TypeToken<Map<String, Any?>>() {}.type
                 )
                 _tipoMuerte.value = datos["tipo"] as? String ?: ""
                 _codigoTipoMuerte.value = datos["codigoTipo"] as? String ?: ""
